test: move genset test from dto to yaml folder
equinor/ecalc-internal#291
